google_gemini/ai-powered-travel-agent/api/requirements.txt
fastapi
uvicorn
python-dotenv
google-genai
python-multipart
Pillow
pydantic